Changed about poster to get questionnair data and made all questionaire inputs required
If you try to submit an empty questionnaire you get an alert telling you to fill every field out. Once you do, the questionnaire is submitted and its data is used for the about poster section in the listing details pages. I added a function in the firebase helpers to help get the data. I also change some colors.